requirements.txt
colorama>=0.4.4
paramiko>=2.7.2
PySocks>=1.7.1
requests>=2.28.1 